Move URL tests from utils tests to URL tests.
